reorg code, refactor with some ES6 styling
