+++ README.md
@@ -2,18 +2,13 @@
 
 **Quiz App**
 
-Originally a single js file which was later split into three modules as an assignment.
+Simple quiz application
 
-**Webpack** - mvc.js is defined as the 'entry' location for webpack. mvc.js assisgns a var to each 'require()' function that accepts the question and variables.js files.
+#### Installation
 
-Each of the variables.js and questions.js requires the use of a 'modules.export' at the end of each file in order to allow mvc.js to use the data on the at file.
+* `npm install`
 
-**File Descriptions**
+#### Usage
 
-**mvc.js** - contains the MVC pattern used to control quiz flow, generate views, and define how quiz is played.
-
-**questions.js** - contains the QUESTIONS array holding all question objects.
-
-**variables.js** - contains the variables identifying all the view elements in jQuery.
-
-**index.html** - placeholder for classes and IDs.
+* `npm start` - builds and watches changes to your application
+* `npm build` - builds app to `build` directory, ready to deploy
