+++ index.html
@@ -21,5 +21,4 @@
     <button type="button" class="restart-button">Restart</button>
 </div>
 
-<script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/2.2.4/jquery.min.js"></script>
-<script src="./js/app.js"></script>
+<script src="./build/bundle.js"></script>
