+++ js/data/viewElements.js
@@ -1,0 +1,15 @@
+const $ = require('jquery');
+
+// View items
+
+module.exports = {
+  questionsPageElement: $('.questions-page'),
+  questionCurrentElement: $('.question-current'),
+  questionsTotalElement: $('.questions-total'),
+  questionElement: $('.question'),
+  answersElement: $('.answers'),
+  resultsPageElement: $('.results-page'),
+  scoreElement: $('.score'),
+  restartButtonElement: $('.restart-button')
+};
+
