--- js/viewElements.js
@@ -1,14 +1,0 @@
-// View items
-
-const viewElements = {
-questionsPageElement: $('.questions-page'),
-questionCurrentElement: $('.question-current'),
-questionsTotalElement: $('.questions-total'),
-questionElement: $('.question'),
-answersElement: $('.answers'),
-resultsPageElement: $('.results-page'),
-scoreElement: $('.score'),
-restartButtonElement: $('.restart-button')
-};
-
-module.exports = viewElements;
