+++ package.json
@@ -1,0 +1,29 @@
+{
+  "name": "quiz_app_mvc",
+  "version": "1.0.0",
+  "description": "**Quiz App**",
+  "main": "./js/app.js",
+  "scripts": {
+    "build": "webpack",
+    "start": "webpack --progress --colors --watch"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/cbilliau/Quiz_App_MVC.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/cbilliau/Quiz_App_MVC/issues"
+  },
+  "homepage": "https://github.com/cbilliau/Quiz_App_MVC#readme",
+  "devDependencies": {
+    "babel-core": "^6.10.4",
+    "babel-loader": "^6.2.4",
+    "babel-preset-es2015": "^6.9.0",
+    "webpack": "^1.13.1"
+  },
+  "dependencies": {
+    "jquery": "^3.0.0"
+  }
+}
